Move decorator to utils
